+++ src/FsToolkit.ErrorHandling.IcedTasks/CancellableValueTaskResultCE.fs
@@ -366,8 +366,8 @@ module CancellableValueTaskResult =
     /// <summary>
     /// Maps the values of an <c>CancellableValueTaskResult</c> to a new <c>CancellableValueTaskResult</c>  using the provided functions.
     /// </summary>
-    /// <param name="onOk">The function to apply to the 'ok' value of the input <c>CancellableValueTaskResult</c>.</param>
-    /// <param name="onError">The function to apply to the 'error' value of the input <c>CancellableValueTaskResult</c>.</param>
+    /// <param name="onOk">The function to apply to the 'ok' value of the input.</param>
+    /// <param name="onError">The function to apply to the 'error' value of the input.</param>
     /// <param name="input">The input <c>CancellableValueTaskResult</c> to map.</param>
     /// <returns>A new <c>CancellableValueTaskResult</c> with the mapped values.</returns>
     let inline eitherMap